+++ src/main/java/cn/popo/news/core/controller/api/CommentController.java
@@ -1,6 +1,7 @@
 package cn.popo.news.core.controller.api;
 
 import cn.popo.news.common.utils.KeyWordFilter;
+import cn.popo.news.common.utils.UserSessionUtil;
 import cn.popo.news.core.entity.common.CommentPraise;
 import cn.popo.news.core.entity.form.CommentForm;
 import cn.popo.news.core.entity.form.CommentReportForm;
@@ -11,6 +12,8 @@ import cn.popo.news.core.vo.ResultVO;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
 import javax.validation.Valid;
 import java.util.HashMap;
 import java.util.Map;
@@ -23,6 +26,8 @@ public class CommentController {
     private AgoCommentService agoCommentService;
     @Autowired
     private CommentPraiseRepository commentPraiseRepository;
+    @Autowired
+    private UserSessionUtil userSessionUtil;
 
     /**
      * @param userId commentId
@@ -32,7 +37,13 @@ public class CommentController {
     @PostMapping("/comment/praise")
     public ResultVO<Map<String,Object>> articleCommentPraise(
                                                         @RequestParam(value = "commentId") String commentId,
-                                                       @RequestParam(value = "userId") String userId){
+                                                       @RequestParam(value = "userId") String userId,
+                                                        HttpServletRequest request,
+                                                        HttpServletResponse response){
+
+        if (!userSessionUtil.verifyLoginStatus(request,response)){
+            return ResultVOUtil.error(3,"用户失效");
+        }
         CommentPraise commentPraise = commentPraiseRepository.findAllByUidAndCommentId(userId,commentId);
         if(commentPraise!=null){
             return ResultVOUtil.error(100,"已赞");
@@ -49,7 +60,12 @@ public class CommentController {
      */
     @PostMapping("/comment/praise/delete")
     public ResultVO<Map<String,Object>> articleDeletePraise(@RequestParam(value = "commentId") String commentId,
-                                                            @RequestParam(value = "commentPraiseId") Integer commentPraiseId){
+                                                            @RequestParam(value = "commentPraiseId") Integer commentPraiseId,
+                                                            HttpServletRequest request,
+                                                            HttpServletResponse response){
+        if (!userSessionUtil.verifyLoginStatus(request,response)){
+            return ResultVOUtil.error(3,"用户失效");
+        }
         agoCommentService.deleteCommentPraise(commentPraiseId,commentId);
         Map<String,Object> map  = new HashMap<>();
         return ResultVOUtil.success(map);
@@ -61,7 +77,12 @@ public class CommentController {
      * @desc 评论上传
      */
     @PostMapping("/comment/save")
-    public ResultVO<Map<String,Object>> commentSave(@Valid CommentForm commentForm){
+    public ResultVO<Map<String,Object>> commentSave(@Valid CommentForm commentForm,
+                                                    HttpServletRequest request,
+                                                    HttpServletResponse response){
+        if (!userSessionUtil.verifyLoginStatus(request,response)){
+            return ResultVOUtil.error(3,"用户失效");
+        }
         String content = commentForm.getCommentInfo();
         if (!KeyWordFilter.checkWords(content).equals("")){
             return ResultVOUtil.error(100,"评论内容违规："+KeyWordFilter.checkWords(content));
@@ -77,7 +98,12 @@ public class CommentController {
      * @desc 评论举报上传
      */
     @PostMapping("/comment/report")
-    public ResultVO<Map<String,Object>> commentReportSave(@Valid CommentReportForm commentReportForm){
+    public ResultVO<Map<String,Object>> commentReportSave(@Valid CommentReportForm commentReportForm,
+                                                          HttpServletRequest request,
+                                                          HttpServletResponse response){
+        if (!userSessionUtil.verifyLoginStatus(request,response)){
+            return ResultVOUtil.error(3,"用户失效");
+        }
         String content = commentReportForm.getContent();
         if (!KeyWordFilter.checkWords(content).equals("")){
             return ResultVOUtil.error(100,"举报内容违规："+KeyWordFilter.checkWords(content));
